// src/models/assign-warehouse1.model.ts
export interface IAssignWarehouse1 {
    id : number | string;
    nombre: string | null;
    apellidoPaterno: string | null;
    apellidoMaterno: string | null;
    kg_frutas_verduras: number | string;
    kg_pan: number | string;
    kg_abarrotes: number | string;
    kg_no_comestibles: number | string;
}